+++ mfhdf/libsrc/CMakeLists.txt
@@ -31,7 +31,7 @@ set (HDF4_MFHDF_LIBSRC_CSRCS
     ${HDF4_MFHDF_LIBSRC_SOURCE_DIR}/sharray.c
     ${HDF4_MFHDF_LIBSRC_SOURCE_DIR}/string.c
     ${HDF4_MFHDF_LIBSRC_SOURCE_DIR}/var.c
-    ${HDF4_MFHDF_LIBSRC_SOURCE_DIR}/xdr.c
+    ${HDF4_MFHDF_LIBSRC_SOURCE_DIR}/h4_xdr.c
 )
 
 set (HDF4_MFHDF_LIBSRC_INSTALL_CHDRS
@@ -46,7 +46,7 @@ set (HDF4_MFHDF_LIBSRC_CHDRS
     ${HDF4_MFHDF_LIBSRC_INSTALL_CHDRS}
     ${HDF4_MFHDF_LIBSRC_SOURCE_DIR}/error.h
     ${HDF4_MFHDF_LIBSRC_SOURCE_DIR}/mfprivate.h
-    ${HDF4_MFHDF_LIBSRC_SOURCE_DIR}/xdr.h
+    ${HDF4_MFHDF_LIBSRC_SOURCE_DIR}/h4_xdr.h
 )
 
 if (HDF4_ENABLE_NETCDF)
